server/requirements.txt
firebase_admin
paho-mqtt
sklearn